+++ crates/ludtwig-parser/CHANGELOG.md
@@ -1,6 +1,12 @@
 # NEXT-VERSION
 
+# v0.6.0
+
 - [#50](https://github.com/MalteJanz/ludtwig/pull/50) Expose `Preorder` type for AST traversal
+- [#109](https://github.com/MalteJanz/ludtwig/pull/109) Added parser documentation and some more AST utility method
+  implementations
+- [#116](https://github.com/MalteJanz/ludtwig/pull/116) Fixed parsing of raw text HTML elements like inline script and
+  style tags
 
 # v0.5.3